reconfigured services
