+++ StudentCompass.Server/nlog.config
@@ -8,7 +8,18 @@
     <target xsi:type="Console" name="console" layout="${longdate} ${level:uppercase=true} ${logger} ${message} ${exception}" />
 
     <!-- File target -->
-    <target xsi:type="File" name="logfile" fileName="logs/app.log" layout="${longdate} ${level:uppercase=true} ${logger} ${message} ${exception}" />
+    <target xsi:type="File" name="allFile" fileName="${basedir}\Logs\${shortdate}.json"
+            archiveAboveSize="1048576"
+            archiveFileName="${basedir}/logs/${shortdate}.{#}.json"
+            archiveNumbering="Sequence">
+      <layout xsi:type="JsonLayout">
+        <attribute name="time" layout="${longdate}" />
+        <attribute name="level" layout="${uppercase:${level}}"/>
+        <attribute name="logger" layout="${logger}"/>
+        <attribute name="Line" layout="${callsite-linenumber}" />
+        <attribute name="message" layout="${message} ${exception:format=tostring,StackTrace}" />
+      </layout>
+    </target>
   </targets>
 
   <rules>
